+++ src/jme3utilities/nifty/test/StarfieldState.java
@@ -264,13 +264,12 @@ public class StarfieldState extends SimpleAppState {
         rootNode.attachChild(objects);
         /*
          * The moving objects are small white spheres, distributed 
-         * pseudo-randomly  throughout a cubical zone centered on the origin
+         * pseudo-randomly throughout a cubical zone centered on the origin
          */
         NameGenerator names = new NameGenerator();
         Material white = MyAsset.createUnshadedMaterial(
                 assetManager, objectColor);
-        float objectRadius = 0.01f * zoneRadius
-                / MyMath.cubeRoot((float) numObjects);
+        float objectRadius = 0.01f * zoneRadius / MyMath.cubeRoot(numObjects);
         Mesh objectMesh = new Sphere(3, 10, objectRadius);
 
         for (int i = 0; i < numObjects; i++) {
